figure 11 edits (#675)
* figure 11 edits

* more

* edits

* fix

* should be error free

* finished?
