+++ manuscript/03.results.md
@@ -62,14 +62,16 @@ As an application of our model, we employed phenotypic measurements of G1 and S/
 
 ### Identified subpopulations own unique properties in drug-treated experimental data
 
-Realizing the most likely number of distinct subpopulations in a data is most valuable when it is accompanied with the quality of each subpopulation. To find out what information each state holds and learning about unique properties of each subpopulation, we plotted the observed measurements for each state, separately. The following two sections discuss our findings for lapatinib and gemcitabine-treated data.
+Realizing the number of distinct subpopulations in a data is most valuable when it is accompanied with the information about specifications of each subpopulation. To find out what information each state holds and learning about unique properties of each subpopulation, we plotted the observed measurements for each state, separately. The following two sections discuss our findings for lapatinib and gemcitabine-treated data.
 
 #### tHMM infers fairly stable subpopulations for lapatinib-treated data
 
-To visualize the properties of the states that the model has identified, we plotted the distribution of average phase-specific time duration for each state and each condition. Figs. {@fig:emissionsLPT}a-d show that treatment with lapatinib leads to three fairly stable states (Fig. @fig:emissionsLPT e) and that infers extension of G1 phase mostly captured in cells of state 0.
+To visualize the properties of the states that the model has identified, we plotted the distribution of average phase-specific time duration for each state and condition. Figs. {@fig:emissionsLPT}a-d show that treatment with lapatinib leads to three fairly stable states (Fig. @fig:emissionsLPT i) and that infers extension of G1 phase mostly captured in cells of state 1 and 2 where we see cells of state 2 in G1 are almost depleted. As expected, lapatinib only affects G1 phase and leaves a very slight influence on G2 phase. 
 
-![**State-specific emissions of the lapatinib-treated data.**](./output/figure11.svg){#fig:emissionsLPT}
+![**State-specific emissions of the lapatinib-treated data. (a)-(d) The distribution of phase lengths for the 3 recognized states in all conditions. (e)-(f) Phase progression rate in units of 1/[hr] for different conditions of the states. (g)-(h) Division rate for G1 phase increases for higher concentrations. (i) State transition probabilities diagram. **](./output/figure11.svg){#fig:emissionsLPT}
 
 #### Gemcitabine treatment leaves stables and cycle states in the cell population
 
-Gemcitabine treatment shown in Figs. {@fig:emissionsLPT}e-h) concludes G2 phase extension as a result of increasing the concentration. There appears to be two stable states and two cycles. 
+Gemcitabine treatment shown in Figs. {@fig:emissionsGMC}a-d) concludes G2 phase extension as a result of increasing the concentration seen in all states. There appears to be two stable states and two cycles (Fig. @fig:emissionsGMC) i. With higher concentrations, cells of all states spend less time in G1 and cell death is the dominant fate in state 0 and 1. 
+
+![**State-specific emissions of the gemcitabine-treated data. (a)-(d) The distribution of phase lengths for the 4 recognized states in all conditions. (e)-(f) Phase progression rate in units of 1/[hr] for different conditions of the states. (g)-(h) Division rate for each phase increases for higher concentrations in some states. (i) State transition probabilities diagram. **](./output/figure12.svg){#fig:emissionsGMC}
